+++ src/domain.rs
@@ -181,7 +181,7 @@ impl<S: PrimeField, G: Group<S>> EvaluationDomain<S, G> {
             {
                 scope.spawn(move |_scope| {
                     for (a, b) in a.iter_mut().zip(b.iter()) {
-                        a.group_sub_assign(&b);
+                        a.group_sub_assign(b);
                     }
                 });
             }
@@ -268,6 +268,7 @@ fn best_fft<S: PrimeField, T: Group<S>>(a: &mut [T], worker: &Worker, omega: &S,
     }
 }
 
+#[allow(clippy::many_single_char_names)]
 fn serial_fft<S: PrimeField, T: Group<S>>(a: &mut [T], omega: &S, log_n: u32) {
     fn bitreverse(mut n: u32, l: u32) -> u32 {
         let mut r = 0;