+++ src/gadgets/num.rs
@@ -124,15 +124,16 @@ impl<Scalar: PrimeField> AllocatedNum<Scalar> {
             found_one |= b;
             if !found_one {
                 // a_bit should also be false
-                a_bit.map(|e| assert!(!e));
+                if let Some(e) = a_bit {
+                    assert!(!e);
+                }
                 continue;
             }
 
             if b {
                 // This is part of a run of ones. Let's just
                 // allocate the boolean with the expected value.
-                let a_bit =
-                    AllocatedBit::alloc(cs.namespace(|| format!("bit {}", i)), a_bit.clone())?;
+                let a_bit = AllocatedBit::alloc(cs.namespace(|| format!("bit {}", i)), a_bit)?;
                 // ... and add it to the current run of ones.
                 current_run.push(a_bit.clone());
                 result.push(a_bit);
@@ -158,8 +159,8 @@ impl<Scalar: PrimeField> AllocatedNum<Scalar> {
 
                 let a_bit = AllocatedBit::alloc_conditionally(
                     cs.namespace(|| format!("bit {}", i)),
-                    a_bit.clone(),
-                    &last_run.as_ref().expect("char always starts with a one"),
+                    a_bit,
+                    last_run.as_ref().expect("char always starts with a one"),
                 )?;
                 result.push(a_bit);
             }
@@ -574,7 +575,7 @@ mod test {
                 .skip(1)
                 .zip(bits.iter().rev())
             {
-                if let &Boolean::Is(ref a) = a {
+                if let Boolean::Is(ref a) = a {
                     assert_eq!(b, a.get_value().unwrap());
                 } else {
                     unreachable!()