+++ src/gadgets/sha256.rs
@@ -37,7 +37,7 @@ where
     assert_eq!(input.len(), 512);
 
     Ok(
-        sha256_compression_function(&mut cs, &input, &get_sha256_iv())?
+        sha256_compression_function(&mut cs, input, &get_sha256_iv())?
             .into_iter()
             .flat_map(|e| e.into_bits_be())
             .collect(),
@@ -77,6 +77,7 @@ fn get_sha256_iv() -> Vec<UInt32> {
     IV.iter().map(|&v| UInt32::constant(v)).collect()
 }
 
+#[allow(clippy::many_single_char_names)]
 fn sha256_compression_function<Scalar, CS>(
     cs: CS,
     input: &[Boolean],
@@ -286,14 +287,13 @@ mod test {
         let mut input_bits: Vec<_> = (0..512).map(|_| Boolean::Constant(false)).collect();
         input_bits[0] = Boolean::Constant(true);
         let out = sha256_compression_function(&mut cs, &input_bits, &iv).unwrap();
-        let out_bits: Vec<_> = out.into_iter().flat_map(|e| e.into_bits_be()).collect();
+        let mut out = out.into_iter().flat_map(|e| e.into_bits_be());
 
         assert!(cs.is_satisfied());
         assert_eq!(cs.num_constraints(), 0);
 
         let expected = hex!("e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855");
 
-        let mut out = out_bits.into_iter();
         for b in expected.iter() {
             for i in (0..8).rev() {
                 let c = out.next().unwrap().get_value().unwrap();