+++ src/gadgets/uint32.rs
@@ -87,11 +87,15 @@ impl UInt32 {
 
         let mut value = Some(0u32);
         for b in bits {
-            value.as_mut().map(|v| *v <<= 1);
+            if let Some(v) = value.as_mut() {
+                *v <<= 1;
+            }
 
             match b.get_value() {
                 Some(true) => {
-                    value.as_mut().map(|v| *v |= 1);
+                    if let Some(v) = value.as_mut() {
+                        *v |= 1;
+                    }
                 }
                 Some(false) => {}
                 None => {
@@ -120,24 +124,32 @@ impl UInt32 {
 
         let mut value = Some(0u32);
         for b in new_bits.iter().rev() {
-            value.as_mut().map(|v| *v <<= 1);
+            if let Some(v) = value.as_mut() {
+                *v <<= 1
+            };
 
             match *b {
                 Boolean::Constant(b) => {
                     if b {
-                        value.as_mut().map(|v| *v |= 1);
+                        if let Some(v) = value.as_mut() {
+                            *v |= 1;
+                        }
                     }
                 }
                 Boolean::Is(ref b) => match b.get_value() {
                     Some(true) => {
-                        value.as_mut().map(|v| *v |= 1);
+                        if let Some(v) = value.as_mut() {
+                            *v |= 1;
+                        }
                     }
                     Some(false) => {}
                     None => value = None,
                 },
                 Boolean::Not(ref b) => match b.get_value() {
                     Some(false) => {
-                        value.as_mut().map(|v| *v |= 1);
+                        if let Some(v) = value.as_mut() {
+                            *v |= 1;
+                        }
                     }
                     Some(true) => {}
                     None => value = None,
@@ -323,7 +335,9 @@ impl UInt32 {
             // Accumulate the value
             match op.value {
                 Some(val) => {
-                    result_value.as_mut().map(|v| *v += u64::from(val));
+                    if let Some(v) = result_value.as_mut() {
+                        *v += u64::from(val);
+                    }
                 }
                 None => {
                     // If any of our operands have unknown value, we won't
@@ -347,11 +361,11 @@ impl UInt32 {
         // The value of the actual result is modulo 2^32
         let modular_value = result_value.map(|v| v as u32);
 
-        if all_constants && modular_value.is_some() {
+        if let (true, Some(result)) = (all_constants, modular_value) {
             // We can just return a constant, rather than
             // unpacking the result into allocated bits.
 
-            return Ok(UInt32::constant(modular_value.unwrap()));
+            return Ok(UInt32::constant(result));
         }
 
         // Storage area for the resulting bits
@@ -509,7 +523,7 @@ mod test {
                         assert!(b.get_value().unwrap() == (expected & 1 == 1));
                     }
                     Boolean::Not(ref b) => {
-                        assert!(!b.get_value().unwrap() == (expected & 1 == 1));
+                        assert!(b.get_value().unwrap() != (expected & 1 == 1));
                     }
                     Boolean::Constant(b) => {
                         assert!(b == (expected & 1 == 1));
@@ -565,6 +579,7 @@ mod test {
     }
 
     #[test]
+    #[allow(clippy::many_single_char_names)]
     fn test_uint32_addmany() {
         let mut rng = XorShiftRng::from_seed([
             0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06,
@@ -602,7 +617,7 @@ mod test {
                         assert!(b.get_value().unwrap() == (expected & 1 == 1));
                     }
                     Boolean::Not(ref b) => {
-                        assert!(!b.get_value().unwrap() == (expected & 1 == 1));
+                        assert!(b.get_value().unwrap() != (expected & 1 == 1));
                     }
                     Boolean::Constant(_) => unreachable!(),
                 }
@@ -705,14 +720,14 @@ mod test {
 
             for b in r.bits.iter() {
                 match b {
-                    &Boolean::Is(ref b) => {
+                    Boolean::Is(ref b) => {
                         assert!(b.get_value().unwrap() == (expected & 1 == 1));
                     }
-                    &Boolean::Not(ref b) => {
-                        assert!(!b.get_value().unwrap() == (expected & 1 == 1));
+                    Boolean::Not(ref b) => {
+                        assert!(b.get_value().unwrap() != (expected & 1 == 1));
                     }
-                    &Boolean::Constant(b) => {
-                        assert!(b == (expected & 1 == 1));
+                    Boolean::Constant(b) => {
+                        assert!(*b == (expected & 1 == 1));
                     }
                 }
 
@@ -749,14 +764,14 @@ mod test {
 
             for b in r.bits.iter() {
                 match b {
-                    &Boolean::Is(ref b) => {
+                    Boolean::Is(ref b) => {
                         assert!(b.get_value().unwrap() == (expected & 1 == 1));
                     }
-                    &Boolean::Not(ref b) => {
-                        assert!(!b.get_value().unwrap() == (expected & 1 == 1));
+                    Boolean::Not(ref b) => {
+                        assert!(b.get_value().unwrap() != (expected & 1 == 1));
                     }
-                    &Boolean::Constant(b) => {
-                        assert!(b == (expected & 1 == 1));
+                    Boolean::Constant(b) => {
+                        assert!(*b == (expected & 1 == 1));
                     }
                 }