+++ src/multiexp.rs
@@ -19,7 +19,7 @@ use super::SynthesisError;
 pub trait SourceBuilder<G: PrimeCurveAffine>: Send + Sync + 'static + Clone {
     type Source: Source<G>;
 
-    fn new(self) -> Self::Source;
+    fn build(self) -> Self::Source;
 }
 
 /// A source of bases, like an iterator.
@@ -45,7 +45,7 @@ impl<G> AddAssignFromSource for G where G: PrimeCurve {}
 impl<G: PrimeCurveAffine> SourceBuilder<G> for (Arc<Vec<G>>, usize) {
     type Source = (Arc<Vec<G>>, usize);
 
-    fn new(self) -> (Arc<Vec<G>>, usize) {
+    fn build(self) -> (Arc<Vec<G>>, usize) {
         (self.0.clone(), self.1)
     }
 }
@@ -130,6 +130,12 @@ impl<'a> QueryDensity for &'a DensityTracker {
     }
 }
 
+impl Default for DensityTracker {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 impl DensityTracker {
     pub fn new() -> DensityTracker {
         DensityTracker { bv: BitVec::new() }
@@ -173,7 +179,7 @@ where
         let mut acc = G::identity();
 
         // Build a source for the bases
-        let mut bases = bases.new();
+        let mut bases = bases.build();
 
         // Create space for the buckets
         let mut buckets = vec![G::identity(); (1 << c) - 1];
@@ -296,7 +302,6 @@ fn test_with_bls12() {
     use ff::Field;
     use group::{Curve, Group};
     use pairing::Engine;
-    use rand;
 
     const SAMPLES: usize = 1 << 14;
 
@@ -313,7 +318,7 @@ fn test_with_bls12() {
             .collect::<Vec<_>>(),
     );
 
-    let naive: <Bls12 as Engine>::G1 = naive_multiexp(g.clone(), v.clone());
+    let naive: <Bls12 as Engine>::G1 = naive_multiexp(g.clone(), v);
 
     let pool = Worker::new();
     let fast = multiexp(&pool, (g, 0), FullDensity, v_bits).wait().unwrap();